@@ -23,8 +23,6 @@
     "prebuild": "rimraf dist",
     "build": "nest build",
     "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\" \"libs/**/*.ts\"",
-    "start": "nest start",
-    "start:dev": "nest start --watch",
     "start:debug": "nest start --debug --watch",
     "start:prod": "node dist/src/main",
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
@@ -42,46 +40,48 @@
     "seed:make": "npm run knex seed:make -- -x ts"
   },
   "dependencies": {
-    "@nestjs/common": "^7.6.12",
+    "@nestjs/common": "^7.6.15",
     "@nestjs/config": "^0.6.3",
-    "@nestjs/core": "^7.6.12",
-    "@nestjs/platform-express": "^7.6.12",
-    "@squareboat/nest-console": "0.0.3",
+    "@nestjs/core": "^7.6.15",
+    "@nestjs/platform-express": "^7.6.15",
+    "@squareboat/nest-console": "0.0.4",
     "@squareboat/nest-events": "0.0.5",
-    "class-transformer": "^0.3.2",
+    "class-transformer": "^0.4.0",
     "class-validator": "^0.13.1",
     "dotenv": "^8.2.0",
-    "express-rate-limit": "^5.2.5",
-    "helmet": "^4.4.1",
-    "knex": "^0.21.17",
+    "express-rate-limit": "^5.2.6",
+    "helmet": "^4.5.0",
+    "knex": "^0.21.15",
     "moment": "^2.29.1",
     "mysql2": "^2.2.5",
-    "objection": "^2.2.14",
-    "query-string": "^6.14.0",
+    "objection": "^2.2.15",
+    "query-string": "^7.0.0",
     "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
-    "rxjs": "^6.6.3",
+    "rxjs": "^6.6.7",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
-    "@nestjs/cli": "^7.5.4",
-    "@nestjs/schematics": "^7.2.7",
-    "@nestjs/testing": "^7.6.12",
+    "@nestjs/cli": "^7.6.0",
+    "@nestjs/schematics": "^7.3.1",
+    "@nestjs/testing": "^7.6.15",
     "@types/express": "^4.17.11",
-    "@types/jest": "^26.0.20",
-    "@types/node": "^14.14.27",
-    "@types/supertest": "^2.0.8",
-    "@typescript-eslint/eslint-plugin": "^4.15.0",
-    "@typescript-eslint/parser": "^4.15.0",
-    "eslint-config-prettier": "^7.2.0",
+    "@types/jest": "^26.0.22",
+    "@types/knex": "^0.16.1",
+    "@types/lodash": "^4.14.168",
+    "@types/node": "^14.14.41",
+    "@types/supertest": "^2.0.11",
+    "@typescript-eslint/eslint-plugin": "^4.22.0",
+    "@typescript-eslint/parser": "^4.22.0",
+    "eslint-config-prettier": "^8.2.0",
     "eslint-plugin-import": "^2.22.1",
     "prettier": "^2.2.1",
     "supertest": "^6.1.3",
-    "ts-jest": "^26.5.1",
-    "ts-loader": "^8.0.17",
+    "ts-jest": "^26.5.5",
+    "ts-loader": "^8.2.0",
     "ts-node": "^9.1.1",
     "tsconfig-paths": "^3.9.0",
-    "typescript": "^4.1.5"
+    "typescript": "^4.2.4"
   },
   "jest": {
     "moduleFileExtensions": [
