@@ -1,5 +1,6 @@
-export * from './Request';
-export * from './Response';
-export * from './Metadata';
-export * from './Decorator';
-export * from './Explorer';
+export * from './metadata';
+export * from './decorators';
+export * from './explorer';
+export * from './controllers/rest';
+export * from './interfaces';
+export * from './guards';
