@@ -12,8 +12,8 @@ import { catchError, timeout } from 'rxjs/operators';
 export class TimeoutInterceptor implements NestInterceptor {
   intercept(context: ExecutionContext, next: CallHandler): Observable<any> {
     return next.handle().pipe(
-      timeout(5000),
-      catchError(err => {
+      timeout(30000),
+      catchError((err) => {
         if (err instanceof TimeoutError) {
           return throwError(new RequestTimeoutException());
         }
