@@ -7,22 +7,14 @@ import { KNEX_CONNECTION } from './constants';
 import { BaseModel } from './db';
 import { DiscoveryModule } from '@nestjs/core';
 
+BaseModel.knex(Knex(KnexConfig));
+
 @Global()
 @Module({
   imports: [DiscoveryModule],
   providers: [
     ...getProviders(),
-    {
-      provide: KNEX_CONNECTION,
-      useFactory: async () => {
-        BaseModel.knex(Knex(KnexConfig));
-        BaseModel.setModulePaths([
-          // add your module names here
-          'user',
-        ]);
-        return Knex(KnexConfig);
-      },
-    },
+    { provide: KNEX_CONNECTION, useFactory: async () => Knex(KnexConfig) },
   ],
   exports: [BaseValidator],
 })
