@@ -2,19 +2,16 @@ import * as Knex from 'knex';
 import { timestamps } from '../helpers';
 
 export async function up(knex: Knex): Promise<void> {
-    return knex.schema.createTable('users', function (table) {
-        table.bigIncrements('id');
-        table.uuid('uuid').index();
-        table.string('first_name');
-        table.string('last_name').nullable();
-        table
-            .string('username')
-            .nullable()
-            .index();
-        timestamps(knex, table);
-    });
+  return knex.schema.createTable('users', function (table) {
+    table.bigIncrements('id');
+    table.uuid('uuid').index();
+    table.string('first_name');
+    table.string('last_name').nullable();
+    table.string('username').nullable().index();
+    timestamps(knex, table);
+  });
 }
 
 export async function down(knex: Knex): Promise<void> {
-    return knex.schema.dropTableIfExists('users');
+  return knex.schema.dropTableIfExists('users');
 }
